+++ changelog.md
@@ -1,7 +1,11 @@
 # Changelog
 
 ## Upcoming
-* Config options have been sorted better
+* Config options have been sorted and categorized
+  * Stamina regen when locked is now a modifier
+* Changed slowdown mechanic
+  * When the stamina is locked the player is now slowed down by 10%, no more thresholds
+  * When the player is sprinting and stamina is below 25% or below 25, sprinting will be less effective (33%)
 * Fixed stamina not synced when switching dimensions
 * Fixed stamina not locked below 2.5 max health